pyproject.toml
[project]
name = "degiro-portfolio"
version = "0.3.6"
description = "DEGIRO portfolio tracking and visualization application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "yfinance>=0.2.0",
    "plotly>=5.24.0",
    "sqlalchemy>=2.0.0",
    "python-multipart>=0.0.21",
    "twelvedata>=1.2.25",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-playwright>=0.4.0",
    "httpx>=0.27.0",
    "invoke>=2.2.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",
    "myst-parser>=2.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/degiro_portfolio"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
# Run tests in parallel using pytest-xdist
# Master DB created once, then copied for each worker (no API calls during tests)
# Each worker gets its own isolated database and server port
# Use -n auto for maximum parallelism, -n 0 for serial execution
addopts = "-v --tb=short -n 4"

[dependency-groups]
dev = [
    "filelock>=3.20.3",
    "pytest-mock>=3.15.1",
    "pytest-xdist>=3.8.0",
]
